requirements.txt
numpy==1.20.3
Pillow
Keras==2.4.2
tensorflow==2.4.2
Flask
Werkzeug
gunicorn
pyttsx3
gdown